use BusinessException and BusinessMessage in DeleteState() action

+++ WebApplication1/Controllers/StateController.cs
@@ -166,10 +166,24 @@ namespace WebApplication1.Controllers
             _logger.LogInformation("State/DeleteState() action is called");
             try
             {
+                if (id == null)
+                {
+                    _logger.LogError("Passed parameter 'id' is null");
+                    throw new BusinessException("Passed parameter 'id' can not be null");
+                }
                 _logger.LogInformation("Calling StateRepository.Delete() method");
                 _stateRepository.Delete(id);                
                 _logger.LogInformation("Redirecting to Index view");
                 return RedirectToAction(nameof(Index));
+            }
+            catch (BusinessException be)
+            {
+                _logger.LogError(be.Message);
+                //lets now create a suitable message for the user and store it inside a ViewBag (which is a Dynamic Object we can fill it
+                //by whatever we want
+                BusinessMessage bm = new BusinessMessage("Error", be.Message);
+                ViewBag.BusinessMessage = bm;
+                return View();
             }
             catch (Exception E)
             {
